Eliminate simple "useless cast" warnings. (programs)
Running GCC with the "useless-cast" warning enabled produces a number
of warnings.  These changes mostly remove simple casting.

+++ mythtv/programs/mythbackend/autoexpire.cpp
@@ -215,7 +215,7 @@ void AutoExpire::CalcParams()
                 uint64_t maxBitrate = enc->GetMaxBitrate();
                 if (maxBitrate==0)
                     maxBitrate = 19500000LL;
-                thisKBperMin += (((uint64_t)maxBitrate)*((uint64_t)15))>>11;
+                thisKBperMin += (maxBitrate*((uint64_t)15))>>11;
                 LOG(VB_FILE, LOG_INFO, QString("    Cardid %1: max bitrate "
                         "%2 Kb/sec, fsID %3 max is now %4 KB/min")
                         .arg(enc->GetInputID())
